Update channel insights charts to use same colors for channels
